backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
numpy>=1.26.0
django-cors-headers==4.3.0
opencv-python==4.8.1.78
Pillow==10.1.0

